[#92] Fix: 시그널 메세지 읽음 처리 오류 수정 (#143)
* refactor: MatchingStatus 통일화

* refactor: signal_message 엔티티 내 인덱스 설정 추가

* refactor: 시그널/채널 내 쿼리 수정 및 이동

* fix: 메세지 읽음 처리 로직 수정
