+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/repository/projection/RoomWithLastSenderProjection.java
@@ -1,0 +1,5 @@
+package com.hertz.hertz_be.domain.channel.repository.projection;
+
+public interface RoomWithLastSenderProjection {
+    Long getLastSenderId();           // 마지막 메시지 보낸 유저 ID
+}
